+++ src/components/List/listChangelogData.ts
@@ -10,6 +10,13 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "3.0.0",
+    type: "Update",
+    affects: ["Styles", "Accessibility"],
+    notes: ["List <ul> bullet color updated."],
+  },
+  {
     date: "2024-03-14",
     version: "3.0.0",
     type: "Update",